+++ library/Episciences/Data.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Episciences;
-
-class Data extends \Episciences_Paper
-{
-
-}