+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - icon to make it easier to identify the user in revision requests section.
 - script to import volumes and volume metadata from journals translation files into the database.
 - Integration of api and application for the bibliographical references
+- [#374](https://github.com/CCSDForge/episciences/issues/374): Importing orcid and affiliations from Zenodo
+- Compatibility Dataverse
 
 ### Changed
 -  repositories config. is now stored in DB. @see /scr/mysql/alter_metadata_sources.sql